Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
diffusers==0.21.4
Pillow==10.0.1
numpy==1.24.3
scikit-image==0.21.0
python-dotenv==1.0.0